+++ extensions/blocks/send-a-message/editor.js
@@ -3,5 +3,14 @@
  */
 import registerJetpackBlock from '../../shared/register-jetpack-block';
 import { name, settings } from '.';
+import {
+	name as whatsAppBlockName,
+	settings as whatsAppBlockSettings,
+} from './service-blocks/whatsapp';
 
-registerJetpackBlock( name, settings );
+registerJetpackBlock( name, settings, [
+	{
+		name: whatsAppBlockName,
+		settings: whatsAppBlockSettings,
+	},
+] );
